PEP8 fixes for file names
From PEP8[1], Modules should have short, all-lowercase names.
Underscores can be used in the module name if it improves readability.
In this repository, there are many python files that do not meet the
standards for PEP8.

This patch renames all of the the python files within the repository
to meet PEP8 naming conventions.

[1] https://www.python.org/dev/peps/pep-0008/#package-and-module-names