Simple fix for dfu flashing
When attempting to run dfu flash on a platform other than Windows
or RaspberryPi, the script downloads a version of dfu-util for
the ARM architecture.  This causes an error when attempting to
execute the program.

This patch adds a line that checks to see if dfu-util is already
available and skips the download process if found.